feat(slash): add remove watcher btn
- add remove wathcer button
- move slash command controller
